Add base class AndroidTVEntity to AndroidTV
